crowd_streamlink/requirements.txt
twitch-python==0.0.19
obs-websocket-py==0.5.3
requests==2.25.1
